+++ wepool-app/src/screens/Root/Home/RideDisplay.tsx
@@ -1,19 +1,18 @@
 import { useState, useEffect } from "react";
-import { HeaderBar } from "../components/HeaderBar";
-import { StyleSheet, View, ScrollView, Text, Touchable } from "react-native";
-import { RootTabScreenProps } from "../navigation/types";
-import { Header } from "../components/Header";
-import { Oops } from "../components/Oops";
-import { RideCard } from "../components/RideCard";
+import { HeaderBar } from "../../../components/HeaderBar";
+import { StyleSheet, View, ScrollView, Text } from "react-native";
+import { RootTabScreenProps } from "../../../navigation/types";
+import { Header } from "../../../components/Header";
+import { Oops } from "../../../components/Oops";
+import { RideCard } from "../../../components/RideCard";
 import { useQuery } from "@apollo/client";
-import { Button } from "../components/Button";
 
 // queries
-import GetOpenRides from "../queries/GET/RideQueries";
+import GetOpenRides from "../../../queries/GET/RideQueries";
 
-export const MyRides = ({
+export const RideDisplay = ({
   navigation,
-}: RootTabScreenProps<"MyRides">) => {
+}: RootTabScreenProps<"Home">) => {
   /**
    * TODO:
    *  - A varible to know user type will be needed
@@ -48,17 +47,9 @@ export const MyRides = ({
         <HeaderBar/>
       </View>
       <View style={styles.contentContainer}>
-        <View style={{flex: 1, flexDirection: "row", alignItems: 'center'}}>
-            <Header text="My Rides" />
-            <View style= {{flex: 2}}></View>
-            <Button 
-                text="+ Create Ride" 
-                style={[styles.button, {backgroundColor: 'green'}]} 
-                textStyle ={styles.buttonText}
-                onPress={() => navigation.navigate("CreateNewRide")}/>
-        </View>
+        <Header text="Current Rides" />
         <View style={styles.cardsContainer}>
-        {openRides ? (
+            {/* {openRides ? (
               <ScrollView>
                   {openRides.map((ride) => {
                     return (
@@ -66,7 +57,10 @@ export const MyRides = ({
                       <RideCard
                           date="20 Apr"
                           time="08:00"
-                          ride={ride}
+                          start_loc={ride.driver?.street}
+                          final_loc={ride.driver?.company?.street}
+                          driverName={ride.driver?.fname}
+                          status={ride.status}
                           handleOnPressEdit={handleOnPressEdit}
                       />
                       </View>
@@ -75,7 +69,7 @@ export const MyRides = ({
               </ScrollView>
             ) : 
             <Oops/>
-            }
+            } */}
         </View>
       </View>
     </View>
@@ -108,17 +102,4 @@ const styles = StyleSheet.create({
     height: 150,
     marginVertical: 10,
   },
-  button: {
-    flex: 1,
-    // width: 45,
-    height: '50%',
-    alignItems: "center",
-    justifyContent: "center",
-    borderRadius: 15,
-    margin: 5,
-    },
-  buttonText: {
-        color: 'white', 
-        fontSize: 18,
-    },
 });
