+++ wepool-app/src/screens/Root/Profile/Profile.tsx
@@ -1,7 +1,7 @@
 import { useState, useEffect } from 'react'
 import { StyleSheet, View, Modal, Text } from 'react-native'
 import { FontAwesome } from '@expo/vector-icons';
-import { useThemeColors } from "../hooks/useThemeColors";
+import { useThemeColors } from "../../../hooks/useThemeColors";
 import { useQuery, gql } from '@apollo/client';
 
 /**
@@ -10,7 +10,7 @@ import { useQuery, gql } from '@apollo/client';
 import { Divider } from '@rneui/themed';            
 
 // queries
-import GetUser from '../queries/GET/UserQueries'
+import GetUser from '../../../queries/GET/UserQueries'
 
 export const Profile = () => {
